refactor(api): home request schemas with modules
